Day 3/RAG_project/requirements.txt
streamlit
google-generativeai
PyPDF2
faiss-cpu
sentence-transformers
